feat: file management is done

+++ controller/file.go
@@ -1,78 +1,91 @@
 package controller
 
 import (
-	"encoding/json"
 	"fmt"
 	"gin-template/common"
 	"gin-template/model"
 	"github.com/gin-gonic/gin"
 	"net/http"
-	"os"
 	"path/filepath"
+	"strconv"
 	"strings"
 	"time"
 )
 
-type FileDeleteRequest struct {
-	Id   int
-	Link string
-	//Token string
+func GetAllFiles(c *gin.Context) {
+	files, err := model.GetAllFiles()
+	if err != nil {
+		c.JSON(http.StatusOK, gin.H{
+			"success": false,
+			"message": err.Error(),
+		})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{
+		"success": true,
+		"message": "",
+		"data":    files,
+	})
+	return
+}
+
+func SearchFiles(c *gin.Context) {
+	keyword := c.Query("keyword")
+	files, err := model.SearchFiles(keyword)
+	if err != nil {
+		c.JSON(http.StatusOK, gin.H{
+			"success": false,
+			"message": err.Error(),
+		})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{
+		"success": true,
+		"message": "",
+		"data":    files,
+	})
+	return
 }
 
 func UploadFile(c *gin.Context) {
+	form, err := c.MultipartForm()
+	if err != nil {
+		c.JSON(http.StatusOK, gin.H{
+			"success": false,
+			"message": err.Error(),
+		})
+		return
+	}
 	uploadPath := common.UploadPath
-	//saveToDatabase := true
-	//path := c.PostForm("path")
-	//if path != "" { // Upload to explorer's path
-	//	uploadPath = filepath.Join(common.ExplorerRootPath, path)
-	//	if !strings.HasPrefix(uploadPath, common.ExplorerRootPath) {
-	//		// In this case the given path is not valid, so we reset it to ExplorerRootPath.
-	//		uploadPath = common.ExplorerRootPath
-	//	}
-	//	saveToDatabase = false
-	//}
-
 	description := c.PostForm("description")
 	if description == "" {
 		description = "无描述信息"
 	}
 	uploader := c.GetString("username")
 	if uploader == "" {
-		uploader = "匿名用户"
+		uploader = "访客用户"
 	}
+	uploaderId := c.GetInt("id")
 	currentTime := time.Now().Format("2006-01-02 15:04:05")
-	form, err := c.MultipartForm()
-	if err != nil {
-		c.String(http.StatusBadRequest, fmt.Sprintf("get form err: %s", err.Error()))
-		return
-	}
 	files := form.File["file"]
 	for _, file := range files {
-		// In case someone wants to upload to other folders.
 		filename := filepath.Base(file.Filename)
-		link := filename
-		savePath := filepath.Join(uploadPath, filename)
-		if _, err := os.Stat(savePath); err == nil {
-			// File already existed.
-			t := time.Now()
-			timestamp := t.Format("_2006-01-02_15-04-05")
-			ext := filepath.Ext(filename)
-			if ext == "" {
-				link += timestamp
-			} else {
-				link = filename[:len(filename)-len(ext)] + timestamp + ext
-			}
-			savePath = filepath.Join(uploadPath, link)
-		}
+		ext := filepath.Ext(filename)
+		link := common.GetUUID() + ext
+		savePath := filepath.Join(uploadPath, link)
 		if err := c.SaveUploadedFile(file, savePath); err != nil {
-			c.String(http.StatusBadRequest, fmt.Sprintf("upload file err: %s", err.Error()))
+			c.JSON(http.StatusOK, gin.H{
+				"success": false,
+				"message": err.Error(),
+			})
 			return
 		}
 		// save to database
 		fileObj := &model.File{
 			Description: description,
 			Uploader:    uploader,
-			Time:        currentTime,
+			UploadTime:  currentTime,
+			UploaderId:  uploaderId,
 			Link:        link,
 			Filename:    filename,
 		}
@@ -81,13 +94,17 @@ func UploadFile(c *gin.Context) {
 			_ = fmt.Errorf(err.Error())
 		}
 	}
-	c.Redirect(http.StatusSeeOther, "./")
+	c.JSON(http.StatusOK, gin.H{
+		"success": true,
+		"message": "",
+	})
+	return
 }
 
 func DeleteFile(c *gin.Context) {
-	var deleteRequest FileDeleteRequest
-	err := json.NewDecoder(c.Request.Body).Decode(&deleteRequest)
-	if err != nil {
+	fileIdStr := c.Param("id")
+	fileId, err := strconv.Atoi(fileIdStr)
+	if err != nil || fileId == 0 {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"success": false,
 			"message": "无效的参数",
@@ -96,15 +113,23 @@ func DeleteFile(c *gin.Context) {
 	}
 
 	fileObj := &model.File{
-		Id: deleteRequest.Id,
+		Id: fileId,
 	}
-	model.DB.Where("id = ?", deleteRequest.Id).First(&fileObj)
+	model.DB.Where("id = ?", fileId).First(&fileObj)
+	if fileObj.Link == "" {
+		c.JSON(http.StatusOK, gin.H{
+			"success": false,
+			"message": "文件不存在！",
+		})
+		return
+	}
 	err = fileObj.Delete()
 	if err != nil {
 		c.JSON(http.StatusOK, gin.H{
 			"success": true,
 			"message": err.Error(),
 		})
+		return
 	} else {
 		message := "文件删除成功"
 		c.JSON(http.StatusOK, gin.H{
